Move substr.c to so-1454-4920.  Update streplace.c and move to so-0440-8170.
