+++ src/so-0440-8170/.gitignore
@@ -1,0 +1,1 @@
+streplace
