+++ src/so-0440-8170/makefile
@@ -1,0 +1,11 @@
+# SO 04408170
+
+include ../../etc/soq-head.mk
+
+TFLAGS   = -DTEST
+PROG1    = streplace
+PROGRAMS = ${PROG1}
+
+all: ${PROGRAMS}
+
+include ../../etc/soq-tail.mk
