+++ src/so-1454-4920/.gitignore
@@ -1,0 +1,1 @@
+substr
