+++ src/so-1454-4920/makefile
@@ -1,0 +1,11 @@
+# SO 14544920
+
+include ../../etc/soq-head.mk
+
+TFLAGS   = -DTEST
+PROG1    = substr
+PROGRAMS = ${PROG1}
+
+all: ${PROGRAMS}
+
+include ../../etc/soq-tail.mk
